Modifications to MoinMoin actions from GraphingWiki
