+++ MoinMoin/action/RenamePage.py
@@ -7,6 +7,7 @@
     @copyright: 2002-2004 Michael Reinsch <mr@uue.org>,
                 2006-2007 MoinMoin:ThomasWaldmann,
                 2007 MoinMoin:ReimarBauer
+                2009 Juhani Eronen <exec@iki.fi> (link renaming)
     @license: GNU GPL, see COPYING for details.
 """
 import re
@@ -14,6 +15,9 @@ from MoinMoin import wikiutil
 from MoinMoin.Page import Page
 from MoinMoin.PageEditor import PageEditor
 from MoinMoin.action import ActionBase
+from MoinMoin.parser.text_moin_wiki import Parser
+
+include_re = re.compile('(<<Include\(([^,\n]+)(.*?)\)>>)')
 
 class RenamePage(ActionBase):
     """ Rename page action
@@ -37,6 +41,10 @@ class RenamePage(ActionBase):
             self.rename_redirect = int(self.request.form.get('rename_redirect', '0'))
         except ValueError:
             self.rename_redirect = 0
+        try:
+            self.rename_inlinks = int(self.request.form.get('rename_inlinks', '0'))
+        except ValueError:
+            self.rename_inlinks = 0
 
     def is_allowed(self):
         may = self.request.user.may
@@ -48,6 +56,144 @@ class RenamePage(ActionBase):
             return _('This page is already deleted or was never created!')
         else:
             return None
+
+    def _inlink_rename(self, page, newpagename, oldpagename, comment):
+        rel_newpagename = wikiutil.RelPageName(page, newpagename)
+
+        # The following regexp match functions search for
+        # occurrences of the target page name, determine
+        # if they're absolute, relative or subpage
+        # matches, and replace them accordingly
+        def word_subfun(mo):
+            match = mo.groups()
+
+            if wikiutil.AbsPageName(page, match[1]) == oldpagename:
+                # If the link was relative:
+                if not oldpagename in match[1]:
+                    # If the new page will be a subpage of the
+                    # source, retain relative link. Else, make
+                    # an absolute link.
+                    if (rel_newpagename.startswith('/') or
+                        rel_newpagename.startswith('../')):
+                        return match[1].replace(
+                            wikiutil.RelPageName(page, oldpagename),
+                            rel_newpagename)
+                    else:
+                        return match[1].replace(
+                            wikiutil.RelPageName(page, oldpagename),
+                            newpagename)
+
+                # Else, change absolute link
+                return match[1].replace(oldpagename, newpagename)
+            # No match in this link -> move on
+            else:
+                return match[1]
+
+        def link_subfun(mo):
+            match = mo.groups()
+
+            if wikiutil.AbsPageName(page, match[1]) == oldpagename:
+                # If the link was relative:
+                if not oldpagename in match[0]:
+                    # If the new page will be a subpage of the
+                    # source, retain relative link. Else, make
+                    # an absolute link.
+                    if (rel_newpagename.startswith('/') or
+                        rel_newpagename.startswith('../')):
+                        return match[0].replace(
+                            wikiutil.RelPageName(page, oldpagename),
+                            rel_newpagename)
+                    else:
+                        return match[0].replace(
+                            wikiutil.RelPageName(page, oldpagename),
+                            newpagename)
+
+                # Else, change absolute link
+                return match[0].replace(oldpagename, newpagename)
+            # No match in this link -> move on
+            else:
+                return match[0]
+
+        def include_subfun(mo):
+            match = mo.groups()
+
+            if wikiutil.AbsPageName(page, match[1]) == oldpagename:
+                # If the link was relative:
+                if not oldpagename in match[0]:
+                    # If the new page will be a subpage of the
+                    # source, retain relative link. Else, make
+                    # an absolute link.
+                    if (rel_newpagename.startswith('/') or
+                        rel_newpagename.startswith('../')):
+                        return match[0].replace(
+                            wikiutil.RelPageName(page, oldpagename),
+                            rel_newpagename)
+                    else:
+                        return match[0].replace(
+                            wikiutil.RelPageName(page, oldpagename),
+                            newpagename)
+
+                # Else, change absolute link
+                return match[0].replace(oldpagename, newpagename)
+            # No match in this link -> move on
+            else:
+                return match[0]
+
+        self.page = PageEditor(self.request, page)
+        savetext = self.page.get_raw_body()
+
+        # Must replace both WikiWords and links, as
+        # [[WikiWord]] is a link.
+        word_re = re.compile(Parser.word_rule, re.VERBOSE)
+        savetext = word_re.sub(word_subfun, savetext)
+        link_re = re.compile(Parser.link_rule, re.VERBOSE)
+        savetext = link_re.sub(link_subfun, savetext)
+
+        # Also handle renaming (non-regexped) includes of the page
+        savetext = include_re.sub(include_subfun, savetext)
+
+        _ = self.request.getText
+
+        success = True
+
+        try:
+            msg = self.page.saveText(savetext, 0, comment=comment,
+                                     notify=False)
+        except self.page.Unchanged:
+            msg = _('Error changing links on page %s!') % (self.page.page_name)
+            success = False
+
+        return success, msg
+
+    def rename_inlinks(self, newpagename, comment, msgs):
+        comment = "%s (%s)" % (comment, _("changed links:") +
+                               " %s -> %s" % (self.pagename, newpagename))
+        # List pages that link to the renamed page
+        pages = set()
+        inlinks = self.request.graphdata.get_in(self.pagename)
+        for type in inlinks:
+            pages.update(inlinks[type])
+
+        # Update listed pages
+        for page in pages:
+            # User rights _ARE_ checked here!
+            if not self.request.user.may.write(page):
+                continue
+
+            # If inlink rename of a single page does not work,
+            # continue but make sure to emit a warning
+            success_i, msg = self._inlink_rename(page, newpagename,
+                                                 oldpagename, comment)
+            if not success_i:
+                success = False
+
+            if msg:
+                msgs.append(msg)
+
+        if not success:
+            msgs.append(_(u'Other pages with inlinks renamed successfully.'))
+
+        return success, msgs
 
     def do_action(self):
         """ Rename this page to "pagename" """
@@ -69,6 +215,9 @@ class RenamePage(ActionBase):
             return success, msgs
 
         msgs = [msgs]
+
+        if rename_inlinks:
+            success, msgs = self.rename_inlinks(newpagename, comment, msgs)
 
         if self.show_redirect and self.rename_redirect:
             self.page = PageEditor(self.request, self.pagename)
@@ -147,6 +296,8 @@ class RenamePage(ActionBase):
             'pagename': wikiutil.escape(self.pagename, True),
             'newname_label': _("New name"),
             'comment_label': _("Optional reason for the renaming"),
+            'inlinks_label': _('Rename links to page too?'),
+            'inlinks_checked': ('', 'checked')[self.request.args.get('inlinks_checked', '0') == '1'],
             'buttons_html': buttons_html,
             'options_html': options_html,
             }
@@ -168,6 +319,13 @@ class RenamePage(ActionBase):
         <td class="content">
             <input type="text" name="comment" size="80" maxlength="200">
         </td>
+    </tr>
+    <tr>
+    <td>
+    <dd>
+        %(inlinks_label)s<input type="checkbox" name="rename_inlinks" value="1" %(inlinks_checked)s>
+    </dd>
+    </td>
     </tr>
     <tr>
         <td></td>
